complete refactor and clean up
