+++ src/app/app-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { BoardComponent } from './boards/board/board.component';
 import { CodenamesBoardComponent } from './boards/codenames-board/codenames-board.component';
 import { SequenceBoardComponent } from './boards/sequence-board/sequence-board.component';
 import { HomeComponent } from './home/home.component';
