--- src/app/boards/board/board.component.html
@@ -1,142 +1,0 @@
-<div [ngClass]="{displayFlex: isSequence}">
-  <div id="boardCol" class="mainContainer">
-    <div id="title" class="title">
-      <p id="codeNamesTitle" class="codeNamesTitle" routerLink="/home" (click)="deleteGameFromDb()"><strong><a>{{title}}</a></strong></p>
-    </div>
-
-    <mat-card id="boardTop" class="boardTop">
-      <table class="matTable" mat-table>
-        <tr mat-table-row>
-          <td mat-cell class="score">
-            <span style="color: red;">Red: {{currentGameIdPair.game.redScore}}</span>&nbsp;&nbsp;&nbsp;&nbsp;<span style="color: blue;">Blue: {{currentGameIdPair.game.blueScore}}</span>
-          </td>
-          <td mat-cell class="turnIndicator">
-            <div *ngIf='!isSequence'>
-              <span class="redText" *ngIf="currentGameIdPair.game.isRedTurn">Red's</span>
-              <span class="blueText" *ngIf="currentGameIdPair.game.isBlueTurn">Blue's</span>
-              <span [ngClass]="{redText: currentGameIdPair.game.isRedTurn, blueText: currentGameIdPair.game.isBlueTurn}"> turn</span>
-            </div>
-            <div *ngIf='isSequence'>
-              <span [ngClass]="{redText: currentGameIdPair.game.currentPlayer.teamColor === 'red', blueText: currentGameIdPair.game.currentPlayer.teamColor === 'blue'}"> {{currentGameIdPair.game.currentPlayer.name}}'s turn</span>
-            </div>
-          </td>
-          <td mat-cell class="endTurnButton" (click)="endTurn()">
-            <button mat-raised-button color="primary">End turn</button>
-          </td>
-        </tr>
-      </table>
-    </mat-card>
-
-    <div id="board" [ngClass]="{boardContainerWords: currentGameIdPair.game.mode.toString() === 'CODENAMES_WORDS',
-                                boardContainerPictures: currentGameIdPair.game.mode.toString() === 'CODENAMES_PICTURES',
-                                boardContainerSequence: currentGameIdPair.game.mode.toString() === 'SEQUENCE'}">
-      <div *ngFor="let card of currentGameIdPair.game.cardsForBoard">
-        <mat-card [ngClass]="{cardContainer: !card.selected,
-                              cardContainerSelectedRed: (card.selected || isSpyMaster) && card.color == 'red',
-                              cardContainerSelectedBlue: (card.selected || isSpyMaster) && card.color == 'blue',
-                              cardContainerSelectedBlack: (card.selected || isSpyMaster) && card.color == 'black',
-                              cardContainerSelectedInnocent: (card.selected || isSpyMaster) && card.color == 'gray'}">
-          <div *ngIf="currentGameIdPair.game.mode.toString() === 'CODENAMES_WORDS'" (click)="select(card)" style="margin: 25px">
-            <div *ngIf="card.color == 'black' && card.selected; else regularWordDisplay">
-              <div>
-                <p class="assassinDisplay">BANG!!!</p>
-                <p class="assassinDisplay" [style.color]="currentGameIdPair.game.isRedTurn ? 'cornflowerblue' : 'tomato'">{{currentGameIdPair.game.isRedTurn ? "Blue" : "Red"}} team wins</p>
-              </div>
-            </div>
-
-            <ng-template #regularWordDisplay>
-              <div>
-                {{ card.word }}
-              </div>
-            </ng-template>
-          </div>
-
-          <div *ngIf="currentGameIdPair.game.mode.toString() === 'CODENAMES_PICTURES'" (click)="select(card)" class="cardImg">
-            <div *ngIf="card.color == 'black' && card.selected; else regularImgDisplay">
-              <div class="assassinImgContents">
-                <p>BANG!!!</p>
-                <p [style.color]="currentGameIdPair.game.isRedTurn ? 'cornflowerblue' : 'tomato'">{{currentGameIdPair.game.isRedTurn ? "Blue" : "Red"}} team wins</p>
-              </div>
-            </div>
-
-            <ng-template #regularImgDisplay>
-              <div>
-                <img src="{{card.imgPath}}" width="200px" height="auto"/>
-              </div>
-            </ng-template>
-          </div>
-
-          <div *ngIf="currentGameIdPair.game.mode.toString() === 'SEQUENCE'">
-            <div class="displayFlex">
-              <span (click)="select(card)" [ngClass]="{redSuit: card.suit === Suit.DIAMOND || card.suit === Suit.HEART }">{{ card.displayValue }}</span>
-              <div *ngIf="card.selected">
-                <button id="removeButton" (click)="removeWithOneEyedJack(card, currentGameIdPair.game.cardsForBoard.indexOf(card))">x</button>
-              </div>
-            </div>
-          </div>
-        </mat-card>
-      </div>
-    </div>
-
-    <mat-card class="boardBottom">
-        <mat-slide-toggle *ngIf="currentGameIdPair.game.mode.toString() !== 'SEQUENCE'" class="spyMasterToggle" color="primary" (click)="toggleSpyMaster()">SpyMaster</mat-slide-toggle>
-        <button id="nextGameButton" class="nextGameButton" mat-raised-button color="primary" (click)="nextGame()">Next Game!</button>
-    </mat-card>
-
-    <div *ngIf="currentGameIdPair.game.mode.toString() === 'CODENAMES_PICTURES'">
-      <p class="copyright"><strong>*** Images adapted from Czech Games Edition, <i>Codenames: Pictures</i>. 2016. ***</strong></p>
-    </div>
-
-  </div>
-
-  <div *ngIf="isSequence" id="sequenceCardsCol" class="playerHandContainer">
-    <div class="playerHandTitle" [style.color]="selectedPlayer.teamColor">
-      {{selectedPlayer.name}}'s hand
-    </div>
-    <div id="playerHand" class="playerHandGrid">
-      <div *ngFor="let card of selectedPlayer.cardsInHand">
-        <mat-card [ngClass]="{cardContainer: !card.selected,
-                              cardContainerSelectedRed: card.selected && card.color == 'red',
-                              cardContainerSelectedBlue: card.selected && card.color == 'blue',
-                              cardContainerSelectedBlack: card.selected && card.color == 'black'}">
-          <div class="cardSequence">
-            <div>
-              <span [ngClass]="{redSuit: card.suit === Suit.DIAMOND || card.suit === Suit.HEART }">{{ card.displayValue }}</span>
-            </div>
-          </div>
-        </mat-card>
-      </div>
-    </div>
-
-    <div id="discardPile" *ngIf="currentGameIdPair.game.topCardOnDiscardPile">
-      <h3>DISCARD PILE:</h3>
-      <mat-card class="cardContainer">
-          <div class="cardSequence">
-            <div>
-              <span [ngClass]="{redSuit: topDiscardIsRedSuit()}">{{ currentGameIdPair.game.topCardOnDiscardPile?.displayValue }}</span>
-            </div>
-          </div>
-        </mat-card>
-    </div>
-    <div *ngIf="currentGameIdPair.game.deck === undefined" id="reshuffleButton">
-      <button mat-raised-button (click)="midGameShuffle()">Shuffle</button>
-    </div>
-  </div>
-
-
-  DECK:
-    <div *ngIf="isSequence">
-      <div  *ngFor="let card of currentGameIdPair.game.deck">
-        {{ card.displayValue }}
-      </div>
-  
-      <br/>
-    </div> 
-    
-
-    <!-- DISCARD PILE: 
-    <div  *ngFor="let card of currentGameIdPair.game.discardPile">
-      {{ card.displayValue }}
-    </div> -->
-
-</div>
