+++ src/app/boards/sequence-board/sequence-board.component.css
@@ -51,20 +51,6 @@
     text-align: center; width: auto;
   }
   
-  .boardContainerWords{
-    display:grid;
-    grid-template-columns:repeat(5, 19.5%);
-    grid-gap:10px;
-    grid-auto-flow:row;
-  }
-  
-  .boardContainerPictures{
-    display:grid;
-    grid-template-rows:repeat(4, 24%);
-    grid-gap:10px;
-    grid-auto-flow:column;
-  }
-  
   .boardContainerSequence{
     display:grid;
     grid-template-columns:repeat(10, 9.5%);
@@ -114,36 +100,9 @@
     background-color: #8080ff;
   }
   
-  .cardContainerSelectedInnocent{
-    text-align: center;
-    font-size: 40px;
-    border: 5px solid gray;
-    padding: 5px;
-    background-color: lightgray;
-  }
-  
-  .cardContainerSelectedBlack{
-    text-align: center;
-    font-size: 40px;
-    border: 5px solid black;
-    padding: 5px;
-    background-color: #424241;
-    color: white;
-  }
-  
   .cardContainer:hover{
     cursor: pointer;
     transform: scale(1.05);
-  }
-  
-  .cardImg{
-    max-width: 100%;
-    max-height: 100%;
-  }
-  
-  .cardImg img{
-    max-width: 100%;
-    max-height: 100%;
   }
   
   .boardBottom{
@@ -152,10 +111,6 @@
     margin-top: 10px;
     border: 1px solid black;
     text-align: right;
-  }
-  
-  .spyMasterToggle{
-    margin:0px 10px;
   }
   
   .nextGameButton{
@@ -170,24 +125,8 @@
     color: blue;
   }
   
-  .assassinDisplay{
-    font-size: 20px;
-    margin: 0px;
-  }
-  
-  .assassinImgContents{
-    font-size: 35px;
-    max-height: 100%;
-    max-width: 100%;
-  }
   button:hover{
     transform: scale(1.1);
-  }
-  
-  .copyright{
-    margin: 20px 0px;
-    padding-bottom: 20px;
-    text-align: center;
   }
   
   .redSuit{
@@ -198,6 +137,5 @@
     margin-top: 50px;
     margin-left: 25px;
     margin-right: 25px;
-  
   }
