+++ src/app/boards/sequence-board/sequence-board.component.ts
@@ -5,9 +5,9 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Card } from 'src/app/models/cards/card';
 import { PlayingCard } from 'src/app/models/cards/playing-card';
 import { Face } from 'src/app/models/face';
-import { GameIdPair } from 'src/app/models/game-id-pair';
+import { GameIdPair } from 'src/app/models/game-id-pairs/game-id-pair';
 import { Player } from 'src/app/models/player';
-import { SequenceGameIdPair } from 'src/app/models/sequence-game-id-pair';
+import { SequenceGameIdPair } from 'src/app/models/game-id-pairs/sequence-game-id-pair';
 import { Suit } from 'src/app/models/suit.enum';
 import { PlayerNameDialogComponent } from 'src/app/player-name-dialog/player-name-dialog.component';
 import { SequenceGameService } from 'src/app/services/sequence-game.service';
