--- src/app/models/game-id-pair.spec.ts
@@ -1,10 +1,0 @@
-// import { Card } from '../models/cards/card';
-// import { CodenamesGameContext } from '../models/game-contexts/codenames-game-context';
-// import { GameIdPair } from './game-id-pair';
-// import { GameMode } from './game-mode.enum';
-
-// describe('GameIdPair', () => {
-//   it('should create an instance', () => {
-//     expect(new GameIdPair('abcd1234', new CodenamesGameContext(GameMode.CODENAMES_WORDS, new Array<Card>(), 0, 0, false, false))).toBeTruthy();
-//   });
-// });
