--- src/app/models/sequence.spec.ts
@@ -1,7 +1,0 @@
-import { Sequence } from './sequence';
-
-describe('Sequence', () => {
-  it('should create an instance', () => {
-    expect(new Sequence([])).toBeTruthy();
-  });
-});
