+++ src/app/services/sequence-game.service.ts
@@ -142,39 +142,6 @@ export class SequenceGameService extends GameService {
         }
       }
     } 
-
-    // //Vertical check
-    // for (let i=0; i < sequenceGame.cardsForBoard.length; i++){
-    //   if (this.isValidStartIndexForVerticalCheck(i)){
-    //       let potentialNewSequence = new Sequence([i, i+10, i+20, i+30, i+40]);
-    //       if (this.isFiveInARow(sequenceGame.cardsForBoard, potentialNewSequence, currentPlayerTeamColor) && !this.isExistingSequence(potentialNewSequence, sequenceGame.existingSequences)){
-    //         this.handleNewSequence(sequenceGame, potentialNewSequence, currentPlayerTeamColor);
-    //         return;
-    //     }
-    //   }
-    // }
-
-    // //Diagonal Positive check
-    // for (let i=0; i < sequenceGame.cardsForBoard.length; i++){
-    //   if (this.isValidStartIndexForDiagonalPositiveCheck(i)){
-    //     let potentialNewSequence = new Sequence([i, i+9, i+18, i+27, i+36]);
-    //     if (this.isFiveInARow(sequenceGame.cardsForBoard, potentialNewSequence, currentPlayerTeamColor) && !this.isExistingSequence(potentialNewSequence, sequenceGame.existingSequences)){
-    //       this.handleNewSequence(sequenceGame, potentialNewSequence, currentPlayerTeamColor);
-    //       return;
-    //     }
-    //   }
-    // }
-
-    // //Diagonal Negative check
-    // for (let i=0; i < sequenceGame.cardsForBoard.length; i++){
-    //   if (this.isValidStartIndexForDiagonalNegativeCheck(i)){
-    //     let potentialNewSequence = new Sequence([i, i+11, i+22, i+33, i+44]);
-    //     if (this.isFiveInARow(sequenceGame.cardsForBoard, potentialNewSequence, currentPlayerTeamColor) && !this.isExistingSequence(potentialNewSequence, sequenceGame.existingSequences)){
-    //       this.handleNewSequence(sequenceGame, potentialNewSequence, currentPlayerTeamColor);
-    //       return;
-    //     }
-    //   }
-    // }
   }
 
   private handleNewSequence(sequenceGame: SequenceGameContext, potentialNewSequence: Sequence, currentPlayerTeamColor: String) {
